+++ AudioHelper/Scripts/Core/AudioHelperRuntimeBootstrapper.cs
@@ -9,6 +9,12 @@ namespace AudioHelper.Core
         {
             AudioHelperConfiguration config = AudioHelperConfiguration.Asset;
 
+            if (config is null)
+            {
+                Debug.LogError("AudioHelperRuntimeBootstrapper.BootstrapAudioManager: No Audio Helper Configuration file found, please run: Tools/Strangeman/Initialize Audio Helper");
+                return;
+            }
+
             if (config.ManagerPersistence is AudioManagerPersistence.RuntimeBootstrap)
             {
                 Object.DontDestroyOnLoad(Object.Instantiate(config.AudioManagerPersistencePrefab));